Introduce a capture-by-move periodic timer system
It's confusing; but I thought it was cool.

The secret-sauce is that the callable procedure (TimerThunk) is embedded in an option
so that the C callback can take it by value (`#take()`) even though it only has an
`&mut reference` to it ...

The other trick is that NextInterval is defined as a new-type struct
to allow for `TimerThunk` to be a recursive type.

A TimerThunk is an (optional) procedure which yields a time in milliseconds
until the next procedure is called, along with the next procedure to call.

As you can see from the tests: you can write simple recursive functions
which meet this definition.
I'm wondering if a macro could be used to auto-generate such a recursive
definition from any given `move|| {}` closure.

Why a Thunk? SDL timers are run in a separate thread: Thunk ensures that
the closure used for the periodic timer:

  * Owns its environment
  * That environment can safely be sent to another task (+Send)

AFAIK Thunk is sort of an implementation detail because FnOnce is
not currently object-safe.